requirements.txt
openai-agents
gradio
pydantic
openai>=1.0.0
python-dotenv